fix(build): correct jOOQ forcedTypes and package alignment for FV-3800
- Change forcedTypes mapping for epoch-day fields from INTEGER to BIGINT to
  match EpochDaysConverter (Long <-> LocalDate) and eliminate type inference
  conflicts during jOOQ code generation.
- Fix jOOQ target packageName to use org.yupay.* instead of com.yupay.*,
  resolving IntelliJ classpath inconsistencies and NoClassDefFoundError during
  build.
- Regenerate jOOQ sources under the unified package structure.

+++ fv3800/build.gradle.kts
@@ -144,7 +144,7 @@ jooq {
 
                             // DATE: of epoch day
                             ForcedType().apply {
-                                name = "INTEGER"
+                                name = "BIGINT"
                                 includeExpression = ".*\\.(birthday|.*_date|.*_at)$"
                                 userType = "java.time.LocalDate"
                                 converter = "com.yupay.amasua.fv3800.model.jooq.converters.EpochDaysConverter"
@@ -152,7 +152,7 @@ jooq {
 
                             // UNSCALED DECIMALS: integers, 6 decimals
                             ForcedType().apply {
-                                name = "INTEGER"
+                                name = "BIGINT"
                                 includeExpression = ".*\\.(value_.*|.*_percentage)$"
                                 excludeExpression = ".*\\.value_type"
                                 userType = "java.math.BigDecimal"
